// src/models/furoui5test/cube/Materials.ts
// Code generated by furo protoc-gen-open-models. DO NOT EDIT.
// protoc-gen-open-models version: ????

import type { FieldNode } from "@furo/open-models/dist/index";
import { ENUM } from "@furo/open-models/dist/index";

//  Types of materials
export enum Materials {
  //  leading
  MATERIALS_UNSPECIFIED = "MATERIALS_UNSPECIFIED", // trailing
  //  glass is cool
  MATERIALS_GLASS = "MATERIALS_GLASS",
  //  wood is warm
  MATERIALS_WOOD = "MATERIALS_WOOD", // very warm
  MATERIALS_CERAMICS = "MATERIALS_CERAMICS",
  MATERIALS_METALS = "MATERIALS_METALS",
  MATERIALS_PLASTICS = "MATERIALS_PLASTICS",
  MATERIALS_TEXTILES = "MATERIALS_TEXTILES", // are nice
  MATERIALS_LEATHER = "MATERIALS_LEATHER",
  //  paper beats
  MATERIALS_PAPER = "MATERIALS_PAPER", // scissors
  MATERIALS_RUBBER = "MATERIALS_RUBBER",
}

/**
 * FieldNode for Materials.
 * Used as the element constructor of ARRAY and the value constructor of MAP,
 * which both build their children with `new Constructor()`.
 **/
export class MaterialsENUM extends ENUM<Materials> {
  constructor(initData?: Materials, parent?: FieldNode, parentAttributeName?: string) {
    super(initData, Materials, Materials.MATERIALS_UNSPECIFIED, parent, parentAttributeName);
  }
}